Initial Commit
Content inside of _UE is pre-built example content provided by Unreal Engine and will not be included in the final build. It is suggested to utilize existing blueprints as a learning aid.

Several base folders with empty blueprints have been created as examples of naming convention and to allow for quicker startup.

Remember to connect Unreal to source control (lower right corner of the editor or Tools -> Revision Control -> Connect to Revision Control. It will prompt you for your execution pathway for git when connecting.