Player interaction stuff
Player can press E or the top button of a gamepad to interact with things. Interaction functionality allows for tapped presses or held presses to be interpreted.

BP_InteractableObject provides a base type for objects that have a visual display showing they can be interacted with hovering over them. This functionality currently creates a widget for each interactable object!!! This will be changed so that the widget is created once and moved around to the object the player is near instead, reducing the number of loaded widgets rather than relying on creating and garbage collecting them.

Refactored a few things maybe? idk lol I haven't been pushing often enough.

Reorganized the folder structure. If anyone else was in the project, sorry I broke everything.

The BP_Door is completely broken at the moment lol I'll fix it soon. The purpose of BP_Door is to serve as a generalized door object that can be modified to suit the specific application e.g. a sliding door, a standard swinging door, etc.

BP_Button acts similarly. It allows the player to interact with it which then sends a signal along to an associated object that it should do stuff. The signal is only sent if the player has the requisite "key" item in their inventory. "Key" items are defined in BP_Key which, along with BP_Button inherits from BP_InteractableObject. The associated data structure for BP_Key which determines the actual subset of key types in the game is E_KeyTypes. A basic, standard door would be a prefab containing a BP_Button set to not need a key and an associated BP_Door set to swing open along the customized hinge.