Doors are working now
Also keys are stored in a more sensible format
Prior to this, keys were stored as an array of actors which is very heavy. This has been changed to an array of E_KeyTypes.

Doors can be set to move on their forward, sideways, or vertical axis (locally defined), or any combination of those. This was way more of a headache than I thought it was going to be.