Enemy Patrol setup, wandering
Enemies can now accept a BP_PatrolRoute asset that determines their patrol points. A spline is used to represent an approximation of the enemy's path. However, the individual spline points are used as the destination and as such the spline shape can differ wildly from the enemy's actual path.

Enemies sometimes wander instead of following their patrol. This is intended behavior.

Next: Allow per-enemy modification of dedication to patrol route.