Updates to enemy wandering, patrol agnostics, enemy data management
Enemies can now set on a per-enemy basis:
• A random patrol radius,
• A +/- random patrol point modifier (random between the set value and its inverse), and
• A random chance to wander between 0 and 100, 0 being never wander and 100 being always wander.

If a patrol is set, the enemy will advance to the first patrol and then lock into wandering until it spots the player. It will then advance to its next patrol point before locking into the wandering pattern again. This is intended behavior.

0 is returned instead if the +/- variance would result in a value less than 0.