WorldGrid Reference Correction
WorldGrid material used in level prototyping was referencing a material function (MF_ProcGrid) and a texture (T_GridChecker) contained inside a folder that should not be included in the remote repository.

In prep for removing /_UE/LevelPrototyping and /_UE/FPS_Refs the WorldGrid material being used in the demo level needed to have references to assets in those folders removed.