Basic enemy attack
The behavior tree initiates a task that sends a message to the enemy beginning the attack.
Implementation of the attack is enemy dependent.
The task waits for 5 seconds before falling back to a 'failed' execution if it hasn't received a notification from the enemy.
This 'hang' is intentional - it is to notify the developer that they have forgotten to implement an attack on the enemy.
Enemies that are intended to have no attacks should implement the attack event anyways and send a message back to the task immediately stating success or failure.
This will likely be updated later to allow for early termination of the behavior tree logic based on enemy data retrieved early on using a 'do-once' structure in the tree.