Cleaning out local content
Gitignore is set to ignore anything inside the _UE folder while keeping the folder itself.

Removed large amount of pre-generated content to reduce size of repo.

Utilize the /_UE/ folder to store anything you don't want in the repo:
Remember that if you're going to use any starter content or similar you need to create a copy and move it into the correct area of the project proper. This must be done for any dependencies the asset may have! Example: You want a useful static mesh from /StarterContent. You also need to ensure that any materials and textures it's using are copied over AND that the mesh and material are targeting those local copies!